Backend/requirements.txt
beautifulsoup4==4.12.2
fastapi==0.108.0
fastapi-sessions==0.3.2
postgres==4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pydantic==2.5.3
pydantic_core==2.14.6
pyjson==1.4.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
starlette==0.32.0.post1
urllib3==2.2.1
uvicorn==0.25.0
